refactor(strava): untangle ingestion module's six responsibilities (#702)
Split strava_ingestion/ingestion.py's bundled contracts into focused,
separately-testable modules and move the two analysis concerns out of the
persistence writer, behaviour-preserving:

- persistence.py: the raw Activity/stream writer (upsert, stream fetch, refetch)
- auth.py: token refresh + per-account Redis lock (#597)
- zone_sync.py: ingestion-time HR-zone calibration
- ingestion.py: slimmed to the two batch orchestrators only
- analysis/zones.py: the pure extract_hr_zone_bounds parser (the seam it serves)
- analysis/intervals.merge_preserved_laps: the lap-preservation rule, no longer
  embedded in the writer
- blocks.py: now owns the block-recovery lifecycle (assign_block_guarded +
  reconcile_unassigned_activities), concentrating that policy where it belongs

The package's public re-exports are unchanged; strava_import.py and the affected
tests are rewired to the moved seams. All moved bodies are verbatim (only
_assign_block -> assign_block_guarded and the lap-block extraction). Full suite
green (2217 passed) and independently confirmed behaviour-equivalent.
